+++ services-directions-models/src/main/java/com/mapbox/api/directions/v5/models/Congestion.java
@@ -62,7 +62,7 @@ public abstract class Congestion extends DirectionsJsonObject {
    * This builder can be used to set the values describing the {@link Congestion}.
    */
   @AutoValue.Builder
-  public abstract static class Builder {
+  public abstract static class Builder extends DirectionsJsonObject.Builder<Builder> {
 
     /**
      * Quantitative descriptor of congestion. 0 to 100.