+++ services-directions-models/src/main/java/com/mapbox/api/directions/v5/models/DirectionsWaypoint.java
@@ -111,7 +111,7 @@ public abstract class DirectionsWaypoint extends DirectionsJsonObject {
    * @since 3.0.0
    */
   @AutoValue.Builder
-  public abstract static class Builder {
+  public abstract static class Builder extends DirectionsJsonObject.Builder<Builder> {
 
     /**
      * Provides the way name which the waypoint's coordinate is snapped to.