+++ services-directions-models/src/main/java/com/mapbox/api/directions/v5/models/LegAnnotation.java
@@ -144,7 +144,7 @@ public abstract class LegAnnotation extends DirectionsJsonObject {
    * @since 3.0.0
    */
   @AutoValue.Builder
-  public abstract static class Builder {
+  public abstract static class Builder extends DirectionsJsonObject.Builder<Builder> {
 
     /**
      * The distance, in meters, between each pair of coordinates.