+++ services-directions-models/src/main/java/com/mapbox/api/directions/v5/models/MapboxStreetsV8.java
@@ -76,7 +76,7 @@ public abstract class MapboxStreetsV8 extends DirectionsJsonObject {
    * This builder can be used to set the values describing the {@link MapboxStreetsV8}.
    */
   @AutoValue.Builder
-  public abstract static class Builder {
+  public abstract static class Builder extends DirectionsJsonObject.Builder<Builder> {
 
     /**
      * Class of the road exiting the intersection.