+++ services-directions-models/src/main/java/com/mapbox/api/directions/v5/models/MaxSpeed.java
@@ -106,7 +106,7 @@ public abstract class MaxSpeed extends DirectionsJsonObject {
    * @since 3.0.0
    */
   @AutoValue.Builder
-  public abstract static class Builder {
+  public abstract static class Builder extends DirectionsJsonObject.Builder<Builder> {
 
     /**
      * Number indicating the posted speed limit.