+++ services-directions-models/src/main/java/com/mapbox/api/directions/v5/models/StepManeuver.java
@@ -324,7 +324,7 @@ public abstract class StepManeuver extends DirectionsJsonObject {
    * @since 3.0.0
    */
   @AutoValue.Builder
-  public abstract static class Builder {
+  public abstract static class Builder extends DirectionsJsonObject.Builder<Builder> {
 
     /**
      * The rawLocation as a double array. Once the {@link StepManeuver} object's created, this raw