+++ services-directions-models/src/main/java/com/mapbox/api/directions/v5/models/TollCollection.java
@@ -72,7 +72,7 @@ public abstract class TollCollection extends DirectionsJsonObject {
    * This builder can be used to set the values describing the {@link TollCollection}.
    */
   @AutoValue.Builder
-  public abstract static class Builder {
+  public abstract static class Builder extends DirectionsJsonObject.Builder<Builder> {
 
     /**
      * The type of toll collection point, either `toll_booth` or `toll_gantry`.