+++ services-directions-refresh-models/src/main/java/com/mapbox/api/directionsrefresh/v1/DirectionsRefreshAdapterFactory.java
@@ -1,7 +1,7 @@
 package com.mapbox.api.directionsrefresh.v1;
 
 import com.google.gson.TypeAdapterFactory;
-import com.ryanharter.auto.value.gson.GsonTypeAdapterFactory;
+import com.mapbox.auto.value.gson.GsonTypeAdapterFactory;
 
 /**
  * Required so that AutoValue can generate specific type adapters when needed inside the direction